pyproject.toml
[project]
name = "notfall-termin-abh-muenchen-bot"
version = "2.0.0"
description = "Munich appointment booking notification bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    # Configuration & Validation
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # Database
    "sqlmodel>=0.0.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]
